+++ src/components.d.ts
@@ -187,7 +187,11 @@ export namespace Components {
     /**
     * Plan to provision (slug)
     */
-    'planLabel'?: string;
+    'planId'?: string;
+    /**
+    * Product ID
+    */
+    'productId'?: string;
     /**
     * Product to provision (slug)
     */
@@ -200,10 +204,6 @@ export namespace Components {
     * The label of the resource to provision
     */
     'resourceLabel'?: string;
-    /**
-    * _(hidden)_
-    */
-    'restFetch'?: RestFetch;
   }
   interface ManifoldDataRenameButton {
     'disabled'?: boolean;
@@ -1239,7 +1239,11 @@ declare namespace LocalJSX {
     /**
     * Plan to provision (slug)
     */
-    'planLabel'?: string;
+    'planId'?: string;
+    /**
+    * Product ID
+    */
+    'productId'?: string;
     /**
     * Product to provision (slug)
     */
@@ -1252,10 +1256,6 @@ declare namespace LocalJSX {
     * The label of the resource to provision
     */
     'resourceLabel'?: string;
-    /**
-    * _(hidden)_
-    */
-    'restFetch'?: RestFetch;
   }
   interface ManifoldDataRenameButton {
     'disabled'?: boolean;